+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/HubdleProjectPlugin.kt
@@ -6,6 +6,7 @@ import com.javiersc.hubdle.project.extensions.config._internal.createHubdleConfi
 import com.javiersc.hubdle.project.extensions.java._internal.createHubdleJavaExtensions
 import com.javiersc.hubdle.project.extensions.kotlin._internal.createHubdleKotlinExtensions
 import com.javiersc.hubdle.project.extensions.shared._internal.createHubdleSharedExtensions
+import com.javiersc.hubdle.project.tasks._internal.registerHubdleLifecycleTasks
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.plugins.BasePlugin
@@ -15,6 +16,7 @@ public abstract class HubdleProjectPlugin : Plugin<Project> {
 
     override fun apply(target: Project) {
         target.pluginManager.apply(BasePlugin::class)
+        target.registerHubdleLifecycleTasks()
 
         target.configureHubdleExtensions()
     }