+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/config/install/pre/commits/tasks/InstallTestsPreCommitTask.kt
@@ -1,6 +1,6 @@
 package com.javiersc.hubdle.project.extensions.config.install.pre.commits.tasks
 
-import com.javiersc.gradle.tasks.extensions.namedLazily
+import com.javiersc.hubdle.project.tasks.lifecycle.TestsTask
 import javax.inject.Inject
 import org.gradle.api.Project
 import org.gradle.api.file.ProjectLayout
@@ -9,11 +9,12 @@ import org.gradle.api.provider.Property
 import org.gradle.api.tasks.CacheableTask
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.TaskAction
+import org.gradle.kotlin.dsl.named
 import org.gradle.kotlin.dsl.property
 import org.gradle.kotlin.dsl.register
 
 @CacheableTask
-public abstract class InstallAllTestsPreCommitTask
+public abstract class InstallTestsPreCommitTask
 @Inject
 constructor(
     objects: ObjectFactory,
@@ -25,7 +26,8 @@ constructor(
     }
 
     @Input
-    override val preCommitName: Property<String> = objects.property<String>().convention("allTests")
+    override val preCommitName: Property<String> =
+        objects.property<String>().convention(TestsTask.NAME)
 
     @TaskAction
     override fun install() {
@@ -33,20 +35,19 @@ constructor(
     }
 
     public companion object {
-        public const val name: String = "installAllTestsPreCommit"
+        public const val NAME: String = "installTestsPreCommit"
 
         internal fun register(project: Project) {
 
-            val installAllTestsPreCommitTask =
-                project.tasks.register<InstallAllTestsPreCommitTask>(name)
+            val installTestsPreCommitTask = project.tasks.register<InstallTestsPreCommitTask>(NAME)
 
-            installAllTestsPreCommitTask.configure { task ->
+            installTestsPreCommitTask.configure { task ->
                 task.finalizedBy(WriteFilePreCommitTask.getTask(project))
             }
 
-            project.tasks
-                .namedLazily<InstallPreCommitTask>(InstallPreCommitTask.name)
-                .configureEach { task -> task.dependsOn(installAllTestsPreCommitTask) }
+            project.tasks.named<InstallPreCommitTask>(InstallPreCommitTask.NAME).configure { task ->
+                task.dependsOn(installTestsPreCommitTask)
+            }
         }
     }
 }