+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/config/install/pre/commits/tasks/WriteFilePreCommitTask.kt
@@ -52,13 +52,13 @@ constructor(
     }
 
     public companion object {
-        public const val name: String = "writeFilePreCommit"
+        public const val NAME: String = "writeFilePreCommit"
 
         internal fun register(project: Project) {
-            project.tasks.register<WriteFilePreCommitTask>(name)
+            project.tasks.register<WriteFilePreCommitTask>(NAME)
         }
 
         internal fun getTask(project: Project): TaskProvider<WriteFilePreCommitTask> =
-            project.tasks.named<WriteFilePreCommitTask>(name)
+            project.tasks.named<WriteFilePreCommitTask>(NAME)
     }
 }